fix: topic and author profile design review changes
