+++ packages/article-list/src/article-list-item.web.js
@@ -27,6 +27,7 @@ const ArticleListItem = props => {
     onPress,
     publicationName,
     publishedTime,
+    section,
     shortSummary,
     showImage,
     summary,
@@ -63,7 +64,7 @@ const ArticleListItem = props => {
     headline: () => <ArticleSummaryHeadline headline={headline} />,
     labelProps: {
       title: label,
-      color: colours.functional.primary
+      color: colours.section[section] || colours.section.default
     }
   };
