+++ packages/author-profile/src/styles/index.android.js
@@ -2,13 +2,10 @@ import { StyleSheet } from "react-native";
 import sharedStyles from "./shared";
 
 const styles = StyleSheet.create({
-  button: {
-    ...sharedStyles.button,
-    width: 160
-  },
-  textStyle: {
-    ...sharedStyles.textStyle,
-    paddingTop: 1
+  ...sharedStyles,
+  twitter: {
+    ...sharedStyles.twitter,
+    alignItems: "center"
   }
 });
