+++ packages/author-profile/src/styles/index.web.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from "react-native";
+import { spacing } from "@times-components/styleguide";
+import sharedStyles from "./shared";
+
+const styles = StyleSheet.create({
+  ...sharedStyles,
+  jobTitle: {
+    ...sharedStyles.jobTitle,
+    WebkitFontSmoothing: "antialiased"
+  },
+  twitter: {
+    ...sharedStyles.twitter,
+    paddingBottom: spacing(2),
+    paddingTop: spacing(3)
+  }
+});
+
+export default styles;
