+++ packages/provider-test-tools/fixtures/topic-articles.json
@@ -48,7 +48,7 @@
               },
               "__typename": "Image"
             },
-            "section": null,
+            "section": "world",
             "publicationName": "TIMES",
             "publishedTime": "2015-03-17T18:57:58.000Z",
             "summary": [
